Moved python vps module to hoster.vps
