+++ cloudomate/hoster/vps/crowncloud.py
@@ -2,14 +2,13 @@ import re
 import sys
 from collections import OrderedDict
 
-from bs4 import BeautifulSoup
-from mechanicalsoup import LinkNotFoundError
-
 import cloudomate.gateway.bitpay
-from cloudomate.vps.clientarea import ClientArea
-from cloudomate.vps.solusvm_hoster import SolusvmHoster
-from cloudomate.vps.vpsoption import VpsOption
+from bs4 import BeautifulSoup
+from cloudomate.hoster.vps.solusvm_hoster import SolusvmHoster
+from cloudomate.hoster.vps.clientarea import ClientArea
+from cloudomate.hoster.vps.vpsoption import VpsOption
 from cloudomate.wallet import determine_currency
+from mechanicalsoup import LinkNotFoundError
 
 
 class CrownCloud(SolusvmHoster):
