+++ cloudomate/hoster/vps/legionbox.py
@@ -2,9 +2,9 @@ import itertools
 from collections import OrderedDict
 
 from cloudomate.gateway import coinbase
-from cloudomate.vps.clientarea import ClientArea
-from cloudomate.vps.solusvm_hoster import SolusvmHoster
-from cloudomate.vps.vpsoption import VpsOption
+from cloudomate.hoster.vps.solusvm_hoster import SolusvmHoster
+from cloudomate.hoster.vps.clientarea import ClientArea
+from cloudomate.hoster.vps.vpsoption import VpsOption
 
 
 class LegionBox(SolusvmHoster):
