+++ cloudomate/test/test_clientarea.py
@@ -1,9 +1,8 @@
 import os
 import unittest
 
+from cloudomate.hoster.vps.clientarea import ClientArea
 from mock.mock import MagicMock
-
-from cloudomate.vps.clientarea import ClientArea
 
 
 class TestClientArea(unittest.TestCase):
